+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Deprecate using a validation groups generator service not implementing `ApiPlatform\Core\Bridge\Symfony\Validator\ValidationGroupsGeneratorInterface` (#3346)
 * Subresources: subresource resourceClass can now be defined as a container parameter in XML and Yaml definitions
 * IriConverter: Fix IRI url double encoding - may cause breaking change as some characters no longer encoded in output (#3552)
+* GraphQL: Fix "Resource class cannot be determined." error when a null iterable field is returned (#4092)
 
 ## 2.5.6